feat(#33): content에 fetcher util 적용
